+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# 3.8.18 (February 10, 2026)
+
+## Notes
+- Golang: **v1.24**
+
+### Bug Fixes
+
+- [PR #407](https://github.com/zscaler/zscaler-sdk-go/pull/407) - Fixed `GET` function in ZIA `urlfilteringpolicies`.
+
 # 3.8.17 (February 9, 2026)
 
 ## Notes